Add functions to read the cart from ARM7 or ARM9
cardRead() can be used from either CPU to read from the cartridge (but
the user needs to remember to call sysSetCardOwner() before.

cardReadArm7() lets the ARM9 request the ARM7 to read from the cartridge
and save the data in the provided buffer. This will be used in the
future to allow reading data from the cartridge asynchronously while
the ARM9 does other things.