storage: Use locks to protect FIFO communications
After introducing cothreads, whenever we yield from the ARM9 while
waiting for the ARM7 to reply, it is possible that two threads get their
responses in the wrong order.

Thread 1 sends message 1, it yields while waiting for the response,
thread 2 sends message 2, and response 1 is ready, so it reads it
without yielding, and eventually thread 1 sees response 2.

This PR introduces fifoWaitValueAsynch32() so that not all FIFO
communications need to be asynchronous, only the communications that
explicitly use this function.

It also switches the storage and sound FIFO subsystems to use locks
and the new asynchronous wait function.