fifo: Simplify receive interrupt handler
Only allow nested interrupts in safe parts of the code.

Note that the reason to disable the DISP_YTRIGGER_IRQ interrupt was that
this function wasn't really safe to be nested. The ARM7 sends the touch
screen information using FIFO when Y == 80, using the YTRIGGER IRQ, so
disabling it in the receive interrupt handler was a way to workaround a
possible nested interrupt because of the touch screen data.