licenses: Add SPDX identifiers to all files
All license files have been moved to the "licenses" directory in the
root of the repository for clarity.

Using SPDX headers is only used to get clarity. The years displayed in
the notices are the ones that were displayed in the file before, or an
educated guess based on the git history (it assumes that the git
history is all there is to see) or the original authors of ndslib.

The Zlib and BSD licenses don't require the license text to appear in
every file, just as part of the software distribution.

DLDI is licensed under the BSD-3-Clause, and another small header that
can easily be rewritten.

gl2d doesn't have a clear license specified. It was introduced in commit
355ee14e8d317 ("incorporate gl2d") without it. However, it can be
obtained from here [1], so it has been added to this repository.

[1] https://github.com/DeathCamel57/libgl2d_ds/blob/master/distributable/libnds/gl2d_license.txt