nocash: Fix no$gba debug on ARM9 and add support to ARM7
Every time that nocashWrite() is called, no$gba adds a newline character
after the string. Previously, the string was sent to no$gba character by
character, which added a newline after every character.

This patch buffers up to 120 characters (or until a newline character is
found) and sends them to no$gba in one go.

This patch also adds stderr support to the ARM7 by using the no$gba
debug console. It's enabled by default, but it won't be added to the
binary unless it's actually used.