fatfs: Set the CWD to the path of the NDS file
Previously, when initializing FAT, the default path was set to the root
of the drive that contains the NDS file.

With this patch, fatInit() will also set the current working directory
to the folder that contains the NDS file. This way it behaves more like
it would in a real OS