boxtest: Move some functions to the header
Functions that use floats should be static inline so that the compiler can
optimize the call and preconvert the floats during the compilation. It
can't optimize them if the body of the function has already been built as
part of the library.

In the case of the BoxTextResult(), this function is simple enough that it
should be inlined.