fatfs: Always initialize disk cache in nitroFSInit()
The result of not having the cache enabled is that disk_read() would
fail to allocate cache blocks (they would return NULL pointers, so the
card read functions would refuse to read the cartridge to that buffer.
When FatFs received a buffer without any data, it wouldn't detect the
FAT filesystem and return an error.

Previously, only fatInit() would initialize the disk cache. Normally,
nitroFSInit() tries to initialize the DLDI device and the DSi SD card,
so it needs to call fatInit(), which initializes the disk cache.

However, in some emulators (like no$gba or melonDS) nitroFSInit() would
never call fatInit() because no argv[0] was provided. DeSmuMe always
provides argv[0] so, even though fatInit() failed, the cache was
initialized correctly, which allowed NitroFAT to work. melonDS worked
if DLDI was enabled in the emulator options.

This patch ensures that the cache is always initialized, and that it is
reinitialized without memory leaks if required.