mpu: Disable data cache in DTCM
The priority of the DTCM region is lower than the cacheable main RAM
region, which means that DTCM is actually cached. This shouldn't happen,
so this patch reorganizes the regions so that DTCM has priority over the
cacheable main RAM region, and both have priority over the non-cacheable
main RAM region.