rumble: Return "rumble not detected" in DSi mode
The slot-2 memory region can't be accessed in DSi mode because libnds
doesn't map that region in DSi mode. Any access causes a MPU exception.

Other peripherals check if the code is running on a DSi and return
"peripheral not detected" in that case. This patch makes the rumble
behave the same way.

Also, this code adds comments about the reason why the check is there.