rtc: Deprecate usage of RTC interrupt
It isn't supported on 3DS consoles in DS or DSi modes, so any program
using initClockIRQ() that relies on the RTC won't work on 3DS. Because
of that, it is needed to use alternative ways to update the RTC.

Note that using resyncClock() every frame or every few frames isn't a
good option. My tests on hardware show that it uses about 7.5% of the
CPU time available in a frame. Any game that uses audio, Wi-Fi, or
maximizes usage of the ARM7 in general, won't be able to afford that
7.5%.