sound, timers: unify frequency calculation, properly round to nearest frequency
This is an attempt at resolving https://github.com/blocksds/sdk/issues/42
- of course, some testing would be required. `libxm7` works fine, so the
`SOUND_FREQ` values are sound - but this also impacts `TIMER_FREQ`, and
possibly code size/performance. Would be good to look for some homebrew
that make heavy use of those.

This patch makes both defintions round to the nearest frequency, instead
of rounding up. It also documents TIMER_FREQ better, and adds
TIMER_FREQ_SOUND.

Something to consider: Is rounding to nearest, that is *including
rounding down*, for timers acceptable in the first place? It's good for
sampled audio, but what about other usecases?