console: Generate font from PNG file
Instead of having a preconverted font in the repository, it's better to
convert it at build time so that it can be easily modified by other
people.

A small change has been required in the console graphics loading code
because it mirrored tiles from left to right (it only happened with 1bpp
fonts, other sizes are correct).