slot2: Fix title update function
Reading the X and Y values right after sending the sample command doesn't
work, it is needed to give time to the cartridge to update the values.

Games wait until the top bit of TILT_X_HIGH is 1. Instead of that, so that
the function isn't blocking, this commit makes it so that the sample
command is always sent and you can read the results the next frame.