card: nitroFS: Slot-1 NitroFS reads can be done from ARM7
This commit removed the ability of NitroFS Slot-1 reads to be done from
the ARM7:

Commit 4514e731afff ("nitrofs: remove ARM7 card read support (only
worked with main RAM), fix ARM9 card read support a lot").

However, DLDI or DSi SD reads could still be done from the ARM7.

The result is that games behave differently in emulators than in real
hardware. The filesystem asynchronous loading example found in
``examples/filesystem/async_loading`` would work on real hardware, or
when emulating DLDI/DSi SD, but not when emulating Slot-1 card reads.

This commit reintroduces that support so that users can optionally ask
the ARM7 to do the reads. This isn't important just to save cycles in
the ARM9, but to let multithreading work the same way no matter where
NitroFS is reading files from (Slot-1 card commands, DLDI or DSi SD).

The only exception is NitroFS from Slot-2. Reading from the filesystem
in this case is reduced to a simple ``memcpy()``, so it doesn't make
sense to ask the ARM7 to do it because simply sending the FIFO message
around would have a much higher cost than the copy itself.