system: Add reliable way to detect hardware debugger units
The only way to detect the DSi debugger is to read SCFG_EXT9 and check
if the RAM size is 32 MB. The way to detect the DS debugger is to call
swiIsDebugger(), but this only works with the cache disabled (so before
the MPU has been turned on).

This patch modifies __libnds_mpu_setup() to return a flag that the crt0
can use. The crt0 already sets the value of the variable __dsimode, now
it is also in charge of setting __debugger_unit.

isHwDebugger() is a new function introduced to return the value of
__debugger_unit for developers to use it.