grf: Refactor module
Now there are two sets of functions: grfLoad*() and grfLoad*Ex(). The Ex
version supports metatiles and metamaps, the other version doesn't but it
takes fewer arguments.

Also, both sets of functions have been modified to also return the size
of each component of the GRF file, which wasn't available before. This made
it impossible to determine the size of the tileset of tiled backgrounds,
for example.