videoGL: Optimize some functions that use heavy maths
Using `mulf32()` while compiling functions to Thumb makes the compiler
add calls to `__aeabi_lmul()` once per multiplication. This routine is
quite slow, so the cost of any function that uses `mulf32()` a lot is
quite high. `glRotatef32()` was a particularly good example of this
problem.

Making the compiler emit ARM code for functions that use `mulf32()` is
overall a performance and code footprint improvement. This can be
verified by checking the disassembled versions of the ARM and Thumb
versions of the functions. The ARM versions are both smaller and faster.