videoGL: Move big functions outside of the header
It doesn't make to keep functions like `gluLookAtf32()` in the header to be
inlined every time they are used.

Also, in a similar way as commit 130df8f2ba15 ("videoGL: Optimize some
functions that use heavy maths"), the affected functions use maths heavily,
so it makes sense to build them as ARM functions.

The exception is `glResetMatrixStack()`, which wouldn't benefit from being
built as ARM code.