videoGL: Make glCallList() take a void pointer
In most cases, display lists are treated as memory without any specific
structure, not as a list of 32-bit values. glCallList() expected a u32
pointer, which forced any display list exported by bin2c to be casted to
a void pointer. This is a pretty common scenario that shouldn't require a
cast.

This patch makes the function take a void pointer so that it doesn't matter
how the user wants to treat the data.