keyboard: Refactor keyboard init/end code
If keyboardInit()/keyboardDemoInit() is called while there are graphics
already in the tileset/tilemap VRAM, it will display them for a frame
while it's loading the actual graphics of the keyboard.

This patch makes sure to hide the keyboard background layer while the
graphics are being loaded. It also adds the function keyboardEnd(), so that
the keyboard can be hidden right away without the need to change the value
of scrollSpeed.