gl2d: Switch to uint16_t for UV coordinates
Using "unsigned int" is very wasteful, as there is never a situation where
the coordinates are bigger than 1024, which fits in 2 bytes.

NOTE: This is a breaking change, but I'm doing it because I don't think
people are actually using this due to the lack of documentation on how
to generate packed textures.