videoGL: Refactor and fix glColorTableEXT()
It returned success even when it failed to load palettes.

The old structure of the function was bad, it was too nested. This patch
removes two levels of nesting so that it's easier to reason about the logic
of the function and it's easier to see when the function should return
failure or success.