videoGL: Fix allocation of GL_COMPRESSED textures
One part of the texture needs to be allocated in VRAM_B, and the other part
in either VRAM_A or VRAM_C. However, when only VRAM_B and VRAM_C were
configured as texture VRAM, the function failed to allocate textures.

The levels of nesting in the function have also been reduced so that the
logic is easier to follow.