videoGL: Fail properly with invalid sizes in glTexImage2D()
If a size of 2048, for example, was used as input to glTexSizeToEnum(),
it would return -1, but it would be returned as a char, which is
unsigned in ARM, and it would be assigned to sizeX or sizeY as 255.
There is check in glTexImage2D() to return if invalid values (negative
numbers) are detected. However, 255 is positive, so it didn't return.

With this new entry in the GL_TEXTURE_SIZE_ENUM the casts work
correctly, and glTexSizeToEnum() is also cleaner.