console: Don't use BIOS functions in the console
This is a regression caused by commit e70ec2576779 ("console: clean up
consoleLoadFont()"). It was changed as a way to reduce the code
footprint of the function, and we didn't realize that it would break the
default exception handler. A comment has been added so that we are
careful in the future.