nwram: slot2: video: Add memory compiler barriers
The compiler is allowed to reorder accesses between any address that
isn't marked as volatile. This isn't usually a problem because all I/O
registers are marked as volatile in libnds, so all writes and reads have
their order preserved.

However, if RAM is accessed around the writes to the registers, it is
possible that the compiler reorders the access to the register
(volatile) and RAM (non-volatile).

This can be a problem if that specific register can map memory into the
address space of the CPU (or unmap it). This can happen with:

- VRAM bank control registers: They can change the address of the memory
  bank or make it disappear from the address space of the CPU.
- NWRAM bank control registers: They can set the location of each NWRAM
  bank.
- Slot-2 enable/disable routines: They can make new registers or RAM
  banks appear or disappear.

This patch adds compiler barriers to all known problematic functions.

Note that this can't happen with the videoGL registers. They are all
marked as volatile, so it isn't needed to add a barrier to glFlush().