Add defined shift value to bitmasks
This file had no bitshifts defined for the bitmasks, so it was impossible to shift a bit field back to the zero position using defined constants. 

This is useful for addressing arrays for example (one example is an array of strings for debugging).

There are hacks around it but this is a more proper solution.