console: Allow using console with lower code footprint
This commit exposes consolePrintChar() and it stops using printf() in
consoleClear(). This way, it is possible to use the console without ever
requiring to use printf().

However, con_write() is still required, and that function uses sscanf().
In order to also remove that from the binary, con_write() has been
modified to parse integers by itself.

With this last change, a bug has also been fixed. Previously, "[%d;%dH"
wouldn't work. It's the same as "[%d;%df", so it was handled in the same
case clause, but the sscanf() inside was expecting a 'f', not a 'H'.
Also, there was no error checking to see if the new coordinates were
within the bounds of the console, and they have been added with this
commit.