videoGL: Hide internal videoGL state
The only current reason why all the internal structs are visible is that
glClearColor() and glClearPolyID() access the glGlob pointer. Given that
the functions are normally only called once per frame (if at all) the tiny
cost of not inlining them is a cost worth paying to hide all this state.