keyboard: Actually let users load custom keyboards
The pointer passed to keyboardInit_call() is never NULL because it is
only called by keyboardInit(), which passes either the default keyboard
or the custom keyboard passed by the user.

There is no need to do any special handling here. This code was actually
overwriting the custom keyboard state by the default keyboard data.