keyboard: Refactor to handle const init data properly
The structs passed to keyboardInit() should never be modified by libnds
because they are required to be unmodified if we initialize the keyboard
again. It's okay if the user wants to modify them in their code, but
libnds should only work on a copy of the structs.

When keyboardInit() is called now, it copies the keyboard struct into an
internal runtime struct, and it keeps a pointer to the original
initialization data. That way, whenever the keyboard is hidden and shown
again, the keyboard can be restored to the right state (upper-case or
lower-case, shifted, etc).

In order to make this behaviour more explicit, keyboardInit() now takes
a const pointer to a Keyboard struct instead of a non-const pointer.