console: Remove three fields from the PrintConsole struct
They are only used by the default console of libnds. It's better to
hardcode the values outside of the struct so that users don't have to
worry about them.

This will break code that relies on the variables being there, but they
weren't used anyway, so it was misleading, and it's clearer to remove
them.