console: Remove two more unused PrintConsole fields
The field loadGraphics is simply never used at all.

The field consoleInitialised is only ever set to true, never to false.
This means it's actually quite useless because a regular application may
initialize each individual console several times during the duration of
the program. There is no way to set it to false because there is no
consoleExit() or similar.