console: Small cleanup of palette index setup code
Instead of storing the palette index shifted by 12 bits, store the
palette index. The shift is pretty cheap, even when it is done in a
print loop, and it makes the code of the console a lot clearer.

Previously you were supposed to set a plain palette index when
initializing, but it was immediately converted into a palette index
shifted by 12 bits, which is confusing.