console: Support custom fonts properly
When using 4 BPP custom fonts it wasn't possible to specify the index of
the palette to be used for the custom palette. This patch reworks how
the palette index is assigned.

Also, the `fontCharOffset` didn't work. It was used at the time of
writing tile offsets to the background, but it was ignored when loading
tiles to VRAM. This patch fixes the loading code to take this offset
into account.

In order to use both new options properly, `consoleInitEx()` has been
introduced. This is required to preserve compatibility with programs
that are using `consoleInit()` and don't care about the features that
this patch fixes.