console: Print spaces when characters are out of range
Previously, there were no checks to see if the character to be printed
was in the font charset or not. That meant that if you loaded two fonts,
one after the other, you could print the characters of the second one by
printing characters from the first one.

This behaviour was weird and unhelpful, so now the code checks and it
prints spaces whenever the character is outside of the character set of
the font.