slot2: Fix code that detects DSi debugger RAM
When the code was created the MPU didn't distinguish between retail and
debugger models of DSi, so all the 32 MB of potential RAM could be
accessed without triggering an exception. Also, isHwDebugger() didn't
exist at the time.

Since the creation of this code, the MPU setup code now looks for the
debugger extended RAM and isHwDebugger() has been created to detect
debugger models in a reliable way. However, the MPU now blocks accesses
to the extended RAM if the console isn't a debugger unit (so that it's
easier to detect bugs in your code).

This patch simply switches to using isHwDebugger(). We know that if we
reach this function we're in a DSi because that is checked in the only
place where slot2DetectTWLDebugRam() is used.