sound: Allow sound functions select channels to play sound
Most of the code to support this was already there. Even the messages
had a "channel" field that wasn't used. This patch simply updates the
code so that "channel = -1" means that libnds needs to look for a free
channel, and any other value means that libnds needs to use that channel
for the sound.

While it is very useful for libnds to find free channels, it can also be
very useful to manually select the channel.

The old libnds functions to play sounds will keep working as before, but
new functions have been added to be able to select the channel and
support both cases.