sound: Add sound capture helpers
The sound capture can be used to capture the output of sound channels 0
and 2, or the output of the sound mixer to the left or right speakers.
The captured audio can be saved to RAM or added to the output of sound
channels 1 and 3.

While audio capture isn't as useful as video capture, it's still useful
to have a way for users to take advantage of it without too much effort.

Audio capture channel 0 requires sound channel 1 to be active. The
sample rate of the capture is taken from the sample rate of playback of
channel 1 (so that it can be added to it if required). The same applies
to audio capture channel 1 and sound channel 2.

The helpers in this patch only let the user setup the capture registers.
Sound channels 1 and 3 must be setup by hand. The user may want to play
the captured audio there, or it may want to silence it (by setting the
volume to 0) and only use the values saved to RAM.