chore: Warn about unused resulted values that really should be used
If the developer doesn't check whether FAT or NitroFS have been
initialized correctly, it won't be possible to show a sensible error
message to the user of the application when something fails later. It
is very important to verify that they have been initialized correctly
and to show a message to the user so that they know what happened.