grf: Improve code to load GRF files from the filesystem
Instead of loading the whole file and then extracting every chunk that
the developer has requested, only read the parts of the file that
contains the requested chunks.

For example, this can be useful if the developer includes the same
tileset in the GRF files of multiple maps. This way it is possible to
read the map, and only read the tileset if required.