exceptions: Handle ARM7 exceptions and send them to the ARM9
This patch adds an exception handler that the user needs to enable. When
the exception handler is executed, it saves a crash dump in the IPC shared
memory and sends a FIFO message to the ARM9 to notify it so that it can
print it (the same way it prints ARM9 crash dumps). Note that the FIFO
message is just one packet that is sent without using the FIFO library so
that the process is more robust.