asm: Keep functions with fallthrough in the same section
If the top function is called but the bottom one is never referenced in the
code, the garbage collector of the linker may remove the bottom section
completely. In this kind of situations it is important to keep both
functions in the same section (or to add a jump from the top to the bottom,
but that's less efficient).