console: Implement ARM7 console
The ARM9 allocates a ring buffer in main RAM. This buffer is passed to the
ARM7, which then can start writing text to it. The ARM7 needs to notify
the ARM9 when new text is added to the buffer so that the ARM9 can print
it and empty the buffer.

The system is connected to the current ARM9 console system, so the output
of this console can be any PrintConsole in use by the ARM9. For example, it
is possible to start one console on the top screen and let the ARM9 use it
while there is another console on the bottom screen that is used to print
whatever the ARM7 prints.