console: Add ARM7 console print helpers
puts() has been implemented, as well as a minimal version of printf(). ARM7
binaries don't have a lot of space to play with, so it's a generally a bad
idea to use the regular printf() there.

It's still possible to use snprintf() from libc instead of this minimal
version of printf() if more complex formats are required.