cothread: Check that IRQs are on when calling cothread_yield_irq()
If this function is called with interrupts disabled the code will hang
because the only way it will exit is if the requested interrupt happens.

It isn't required to check the I bit in CPSR because user code never
modifies that, and it is always set to enable interrupts by the global
interrupt handler.

Several checks have been added in other functions to give more
meaningful error messages to developers. It's better to know the name of
the top level function that triggered the assert than to know that the
assert happened in cothread_yield_irq().