audio: Protect DSi-only function to set CDC frequency
This function is in a twl section, but there is nothing protecting it from
being called in DS mode. The DSi mode checks are inside the function, so
the caller, when in DS mode, will call the function when it isn't loaded
(hoping that it checks the DSi mode then).

With this commit the mode is checked before calling code in twl sections.