interrupts: Protect against setting NULL as a handler
The interrupt dispatcher isn't protected against using NULL as an interrupt
handler. It only has code to skip calling the interrupt handler if the
pointer is equal to `irqDummy()`. The code in source/common/interrupts.c
is prepared for this and it sets the handler of each interrupt to
`irqDummy()` when the system is initialized and when `irqClear()` is used.

However, some applications do `irqSet(flags, NULL)` as a way to disable a
handler without turning off the interrupt (for example, so that they can
disable a VBL handler but still use swiWaitForVBlank()`.

This means that `irqSet()` needs to translate any NULL pointer to
`irqDummy()` to protect applications that do this.