bios: Add new swiIntrWaitAUX() for the ARM7
The old prototype of swiIntrWait() doesn't support the extended
interrupts of the ARM7. The prototype of swiIntrWaitAUX() supports the
new flags.

However, the old prototype has been kept because the AUX register is
only present in the ARM7 in DSi consoles. It isn't present in DS
consoles even in the ARM7.