input: Update documentation of keysDownRepeat()
This function was very poorly documented. It has been tested and new
documentation has been added to it.

Also, Some people call functions like keysHeld() and keysUp() multiple
times per frame. In other functions this is okay because they don't have
any side-effect. However, keysDownRepeat() has side-effects, so it needs
to be added to the documentation as a warning.