cothread: Optimize cothread_yield_irq()
Previously the main scheduler loop checked every thread context to see
if it was waiting for interrupts and cleared the interrupt wait flag if
the interrupt had been triggered. It would check every context to see if
all threads were waiting for interrupts, and enter halt state if so.
This added a lot of checks to the main scheduler loop.

The new system keeps a counter with the number of total threads and
total threads waiting for interrupts. That way, the scheduler loop only
needs to check if all threads are waiting for an interrupt and enter
halt mode if so.

There are also new arrays of threads waiting for each interrupt. This
makes it very easy for the interrupt handler to iterate that array and
mark all contexts as "ready", saving cycles.

The only code that becomes significantly slower is the code that deletes
a thread context, but that isn't critical.