math: correct semantics of asm in normalizef32()
Kuratius in the gbadev Discord noticed that this asm statement was
missing a memory input constraint, since it reads from the memory
pointed to by `a`.

I took a look at it, too, and noticed a couple more things. `volatile`
isn't necessary because the statement could be eliminated if its outputs
were never used, and the `&` modifiers aren't necessary because the
inputs aren't read from after the outputs are written to. I also changed
the input's name to match the name of the variable instead of the name
of the register it will be kept in.

None of this affects the generated code with the current toolchain, but
it is a little more correct.