fifo: Remove TODO comment
The value of FIFO_MAX_DATA_BYTES is arbitrary, and it only affects the max
number of bytes that a data message can contain. This limit doesn't affect
any of the queues of the FIFO system, so changing this check can't break
anything else. Also, the define name implies that it is possible to send
packets with a size of FIFO_MAX_DATA_BYTES, so ">=" was incorrect because
it only allowed up to "FIFO_MAX_DATA_BYTES - 1" bytes.