irq: Implement swiIntrWait() and swiIntrWaitAUX() in C
The reason why the BIOS has bugs in those functions is that they are
actually quite tricky to get right. Let's just write them in C so that it's
easier to understand the code and to have it bug-free, and let the compiler
generate the assembly code for us.

After checking the generated code, it seems to be reasonably short and
fast.

This patch also moves the ARM9 functions to ITCM.