input: Fix regression causing keysDownRepeat not report keys on their first press
In a previous optimization commit [1], the behaviour of keysDownRepeat
changed.  The state of held keys is updated in `scanKeys`, which calls
`keysDown` internally to build the mask of held keys, but that function
after the optimization is currently returning the previously cached
values which don't reflect the keys being pressed at the moment of
`scanKeys` being called.

[1] Commit 963f8ea6e991 ("input: Save keysDown() and keysUp() state in
scanKeys()")