argv: Support generating argv structs from the DSi device list
This function is to be used by the ARM7 crt0.

If no argv magic is found (thus the loader didn't support the argv
protocol), and in RAM is present the device list, try to retrieve the
file path from there and manually create an argv struct that can then be
parsed by the arm9 crt0, as if it were provided by a separate launcher.

Since the buffer is small, the path is put right at the end of the argv
struct used internally by libnds, allowing everything to be self
contained.

Also, add the relevant address and structures to the system headers.