card: rebrand CARD_CR1_* as a macro of CARD_SPICNTH_*
This aligns with the renaming of the CARD_CR1 register to REG_AUXSPICNT
as of libnds commit e2a8242acde78c2fa65b467e799c3b96d841af52.

The original name is also a misnomer as this macro is intended to be
applied to REG_AUXSPICNTH (the higher 1 byte of REG_AUXSPICNT), which
was actually named CARD_CR1H previously.