exit: Improve compatibility with broken loaders
Some loaders use SCFG_EXT9 to determine the size of DSi RAM, and they
set the size of MPU regions based on this. For example, if bits 14 and
15 are set to 2, a 16 MB region is mapped at 0xC000000. If they are set
to 3, the full 32 MB are mapped.

This isn't the correct way to determine the size of RAM because they are
R/W bits. The right way to do it is to check if the RAM is mirrored
before setting up the MPU. BlocksDS does it that way, and then it sets
the right value in SCFG_EXT9 and the corresponding MPU region

This is a bug that needs to be fixed in all affected loaders, but it's
unlikely that all affected people will update their loaders, so it's
still worth helping the loader do the right thing by changing the value
of SCFG_EXT9.

This issue was introduced in commit 8086e9aecdbc ("mpu: Set RAM size to
16 MB on retail DSi units"), Wed Jan 31 18:14:11 2024.