nitrofs: Check for new ndstool NitroFS magic value
In order to check if card commands work we need a known value that we
can read and check. However, due to the design of the NDS ROM format it
isn't possible to find a single value in the NDS ROM that is well-known
and that can be accessed after boot. Some values in the ROM header could
be used to check, but any offset under 0x8000 can't be read after boot
using card commands.

This commit adds an 8-byte magic string right after the FAT table
because this table must be present in any NDS ROM that uses NitroFS
(FNT is technically optional). The FAT table is guaranteed to be placed
after offset 0x8000 (because it needs to be readable after boot) so this
is a safe place.