stdio: Improve support of stdout and stderr in unistd.h functions
write(STDOUT_FILENO, ptr, size) and write(STDERR_FILENO, ptr, size) now
work as expected.

isatty() is now implemented. The error values set in errno aren't
compliant, but it would be quite a lot of effort to make it compliant (it
depends on external libraries for socket descriptors, for example).

read(), write(), close(), lseek() and ftruncate() now set errno when an
invalid descriptor out of stdin/stdout/stderr is provided to them instead
of only returning an error value.