cache: Remove all calls to DC_InvalidateRange()
Invalidating the data cache in a memory range is unsafe because there
may be variables on the edges of the buffer that get invalidated by
mistake.  It's only safe to invalidate a buffer if it's aligned to a
cache line and the size is a multiple of a cache line.

The safe option is to flush the memory range instead.

Thankfully, in libnds, wherever we invalidate memory we were also
flushing it beforehand, so the code was safe, but the invalidate was
redundant and it only wasted CPU time. The only exception was
`readFirmwareInternal()`, which has been modified to flush memory
instead of invalidating it.

For more information check the two articles below (particularly the
second one):

[1] https://web.archive.org/web/20210622053504/https://www.coranac.com/2009/05/dma-vs-arm9-fight/
[2] https://web.archive.org/web/20210622053550/https://www.coranac.com/2010/03/dma-vs-arm9-round-2/