cp15: Add assertions to DC_InvalidateRange()
It's only safe to invalidate memory if it's aligned to a cache line and
the size of the buffer is a multiple of a cache line. If not there may
be variables around the memory range that get invalidated and lose their
value by mistake.

For more information check the two articles below (particularly the
second one):

[1] https://web.archive.org/web/20210622053504/https://www.coranac.com/2009/05/dma-vs-arm9-fight/
[2] https://web.archive.org/web/20210622053550/https://www.coranac.com/2010/03/dma-vs-arm9-round-2/