nand: no longer write CID to unrelated memory region as side effect
Due to carrying over code, when the nand code was added, the nandCrypt_Init function was made copy the CID to ìREG_CID', which isn't actually a register, but just a memory region used by the official SDK to store the CID in (for use by other applications). This meant that calling this function actually wrote random data to as side effect to region in main ram (which was documented as not reserved).