libc: implement fsync()
IIRC, someone noticed that `fflush()` does not imply a flush to disk,
and as such, the only guaranteed way to synchronize a file's state with
the file system is to use `fclose()`. Unfortunately, modifying picolibc
to have a separate system call route for file flushes is not trivial,
and making every `fwrite()` force a flush to disk is sub-optimal from a
flash endurance and performance perspective.

As a partial solution to the problem, this implements the `fsync()`
POSIX call, which guarantees file system synchronization on a file
descriptor: `fsync(fileno(fp));`. This at least gives homebrew
developers *some* way of forcing such writes, if their usecase requires
it (such as a log file that is never closed).